+++ Firmware/HTLAB.NET_Arduino_DRSSTC_Interrupter/lib_osc.h
@@ -7,8 +7,8 @@
 
 void osc_timer_init();
 void osc_timer_init_64();
-void osc_timer_enable(byte osc, unsigned int period);
-void osc_timer_disable(byte osc);
-void osc_timer_set(byte osc, unsigned int period);
+void osc_timer_enable(uint8_t osc, uint16_t period);
+void osc_timer_disable(uint8_t osc);
+void osc_timer_set(uint8_t osc, uint16_t period);
 
 #endif
